feat: add middlewares

+++ models/event.go
@@ -12,7 +12,7 @@ type Event struct {
 	Description string    `json:"description" binding:"required"`
 	Location    string    `json:"location" binding:"required"`
 	DateTime    time.Time `json:"date"`
-	UserID      int       `json:"userId"`
+	UserID      int64       `json:"userId"`
 }
 
 func (event *Event) Save() error {
